Remove Trends Fragment
